+++ src/bitfield.hpp
@@ -16,7 +16,7 @@ namespace bitfield
      *  ------------
      *  returns 0011
      */
-	Bitfield Set(Bitfield field, Bitfield flag)
+	inline Bitfield Set(Bitfield field, Bitfield flag)
 	{
 		return field | flag;
 	}
@@ -30,7 +30,7 @@ namespace bitfield
      *  ------------
      *  returns 0001
      */
-	Bitfield Clear(Bitfield field, Bitfield flag)
+	inline Bitfield Clear(Bitfield field, Bitfield flag)
 	{
 		return field ^ flag;
 	}
@@ -46,7 +46,7 @@ namespace bitfield
      * 
      *  returns false   true    false   true
      */
-	bool Has(Bitfield field, Bitfield flag)
+	inline bool Has(Bitfield field, Bitfield flag)
 	{
 		return (field & flag) == flag;
 	}
